requirements.txt
tensorboardX
tensorboard
scikit-learn >= 0.22.2
seqeval
pyahocorasick
scipy
statsmodels
biopython
pandas
pybedtools
sentencepiece==0.1.91
transformers
